+++ src/main/java/domain/record/Records.java
@@ -1,7 +1,8 @@
-package record;
+package domain.record;
 
-import car.Car;
-import car.Cars;
+import domain.car.Car;
+import domain.car.Cars;
+import vo.CarRecord;
 
 import java.util.*;
 import java.util.stream.Collectors;
