+++ src/main/java/domain/strategy/MovingStrategy.java
@@ -1,4 +1,4 @@
-package strategy;
+package domain.strategy;
 
 public interface MovingStrategy {
     boolean movable();
