+++ src/main/java/domain/strategy/RandomStrategy.java
@@ -1,4 +1,4 @@
-package strategy;
+package domain.strategy;
 
 import java.util.Random;
